FOC Project 1
Project 1 is all about biodiversity: the variety of different species that exist in a particular habitat. Biodiversity is important because ecosystems with a wider range of species are typically more productive, and have a greater ability to withstand disturbance. A major challenge with biodiversity is how to measure it: we can typically only observe a sample of a large and complex habitat. Birds are considered a good indicator of a healthy ecosystem as they are highly mobile. If they don't like their current surroundings, they will move somewhere else. Even after deciding which species to observe, we still have the problem of working out how to quantify the diversity we observe.

In this project you will write Python functions to explore different questions about biodiversity, in the context of surveying, analysing and predicting the biodiversity of bird species.